requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
bleach==5.0.1
bottle==0.12.23
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==6.5.0
Deprecated==1.2.13
Django==4.1.3
django-appconf==1.0.5
django-compressor==4.1
django-haystack==3.2.1
django-ipware==4.0.2
django-mdeditor==0.1.20
django-mobile==0.7.0
django-uuslug==2.0.0
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
gevent==22.10.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
jieba==0.42.1
jsonpickle==2.2.0
Markdown==3.4.1
mysqlclient==2.1.1
packaging==22.0
Pillow==9.3.0
pycodestyle==2.10.0
Pygments==2.13.0
PySide6==6.2.4
python-dateutil==2.8.2
python-logstash==0.4.8
python-slugify==6.1.2
pytz==2022.6
raven==6.10.0
rcssmin==1.1.0
redis==4.3.4
requests==2.28.1
rjsmin==1.2.0
shiboken6==6.2.4
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
webencodings==0.5.1
WeRoBot==1.13.1
Whoosh==2.7.4
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.5.2
